优化其它命令行启动逻辑。
